Preload account third-party connections like other route data
Connected accounts were fetched after the account route committed,
which left a Loading state (and could get stuck) on SPA navigations.
Load them in the account route loader and SSR payload so the page
renders with connections already present.
